backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
python-dotenv
langchain
unstructured
langchain-community
google-generativeai
libmagic
SQLAlchemy
langchain-postgres
psycopg[binary,pool]==3.2.9
